requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
moviepy>=2.0.0
Pillow>=10.0.0
soundfile>=0.12.0
numpy>=1.24.0
